Fix signed/unsigned comparison warnings in wxVsnprintf test
These warnings only appeared in build configurations using our own
implementation, i.e. not in the default build.

+++ tests/strings/vsnprintf.cpp
@@ -50,27 +50,27 @@ int r;
 
 #define CMP6(expected, fmt, y, z, w, t)                    \
     r=wxSnprintf(buf, MAX_TEST_LEN, wxT(fmt), y, z, w, t); \
-    CPPUNIT_ASSERT_EQUAL( r, wxStrlen(buf) );          \
+    CPPUNIT_ASSERT_EQUAL( r, (int)wxStrlen(buf) );         \
     ASSERT_STR_EQUAL( wxT(expected), buf );
 
 #define CMP5(expected, fmt, y, z, w)                    \
     r=wxSnprintf(buf, MAX_TEST_LEN, wxT(fmt), y, z, w); \
-    CPPUNIT_ASSERT_EQUAL( r, wxStrlen(buf) );          \
+    CPPUNIT_ASSERT_EQUAL( r, (int)wxStrlen(buf) );      \
     ASSERT_STR_EQUAL( wxT(expected), buf );
 
 #define CMP4(expected, fmt, y, z)                     \
     r=wxSnprintf(buf, MAX_TEST_LEN, wxT(fmt), y, z);  \
-    CPPUNIT_ASSERT_EQUAL( r, wxStrlen(buf) );          \
+    CPPUNIT_ASSERT_EQUAL( r, (int)wxStrlen(buf) );    \
     ASSERT_STR_EQUAL( wxT(expected), buf );
 
 #define CMP3(expected, fmt, y)                        \
     r=wxSnprintf(buf, MAX_TEST_LEN, wxT(fmt), y);     \
-    CPPUNIT_ASSERT_EQUAL( r, wxStrlen(buf) );          \
+    CPPUNIT_ASSERT_EQUAL( r, (int)wxStrlen(buf) );    \
     ASSERT_STR_EQUAL( wxT(expected), buf );
 
 #define CMP2(expected, fmt)                           \
     r=wxSnprintf(buf, MAX_TEST_LEN, wxT(fmt));        \
-    CPPUNIT_ASSERT_EQUAL( r, wxStrlen(buf) );          \
+    CPPUNIT_ASSERT_EQUAL( r, (int)wxStrlen(buf) );    \
     ASSERT_STR_EQUAL( wxT(expected), buf );
 
 // NOTE: this macro is used also with too-small buffers (see Miscellaneous())
@@ -354,7 +354,7 @@ void VsnprintfTestCase::S()
 #define CMP3_UTF8(expected, fmt, arg)                                         \
     CPPUNIT_ASSERT_EQUAL                                                      \
     (                                                                         \
-        wxString::FromUTF8(expected).length(),                                \
+        (int)wxString::FromUTF8(expected).length(),                           \
         wxSnprintf(buf, MAX_TEST_LEN, fmt, wxString::FromUTF8(arg))           \
     );                                                                        \
     CPPUNIT_ASSERT_EQUAL                                                      \
